% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{convergent_evol}
\alias{convergent_evol}
\title{Determines the number of convergently evolved traits}
\usage{
convergent_evol(data = NULL)
}
\arguments{
\item{data}{A morpho object}
}
\value{
A data.frame listing convergent traits, their state, and number of transitions
}
\description{
Identifies which traits have evolved independently multiple times
(convergent evolution) in a morpho object.
}
\examples{
phy <- ape::rtree(10)
morpho_data <- sim.morpho(tree = phy, k = 2, trait.num = 20)
convergent_evol(morpho_data)

}
