% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{get.matrix}
\alias{get.matrix}
\title{Get Specific Morphological Matrix}
\usage{
get.matrix(data, seq)
}
\arguments{
\item{data}{A morpho object}

\item{seq}{The sequence data you want to extract}
}
\value{
A data frame with columns edge (branch number), state, and
hmin (where along the branch the transition occurred)
}
\description{
Return data frame of morphological matrix
}
\examples{
phy <- ape::rtree(10)
morpho_data <- sim.morpho(tree = phy, k = 2, trait.num = 20)
get.matrix(morpho_data, tseq)
}
