% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{get.transitions}
\alias{get.transitions}
\title{Get Tranisition History}
\usage{
get.transitions(data, trait)
}
\arguments{
\item{data}{A morpho object}

\item{trait}{The trait number .}
}
\value{
A data frame with columns edge (branch number), state, and
hmin (where along the branch the transition occurred)
}
\description{
Returns the full transition history for a given trait,
including the root state.
}
\examples{
phy <- ape::rtree(10)
morpho_data <- sim.morpho(tree = phy, k = 2, trait.num = 20)
get.transition(morpho_data, trait = 2)
}
