% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateQ.matrices.r
\name{get_gamma_rates}
\alias{get_gamma_rates}
\title{Get discrete gamma rates}
\usage{
get_gamma_rates(alpha, k)
}
\arguments{
\item{alpha}{Numeric. The shape parameter of the gamma distribution.}

\item{k}{Integer. The number of rate categories.}
}
\value{
Numeric vector of length k representing the discrete gamma rates.
}
\description{
Computes a set of discrete gamma rates for rate variation across sites or characters.
This function is adapted from the \code{phangorn} package.
}
\examples{
get_gamma_rates(alpha = 0.5, k = 4)
}
