% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_morpho.R
\docType{data}
\name{morpho_data}
\alias{morpho_data}
\title{Example morpho dataset}
\format{
A list with components:
\describe{
  \item{sequences}{List of sequences for each taxon}
  \item{trees}{List containing the evolutionary tree}
}
}
\usage{
data(morpho_data)
}
\description{
Small example dataset for testing morpho functions.
}
\keyword{datasets}
