% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.tree.R
\name{reconstruct.matrix}
\alias{reconstruct.matrix}
\title{Get reconstructed matrix}
\usage{
reconstruct.matrix(data)
}
\arguments{
\item{data}{A `morpho` object with fossil data}
}
\description{
This function returns the morphological matrix for tips in the
reconstructed tree.
}
