% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateQ.matrices.r
\name{symmetric.Q.matrix}
\alias{symmetric.Q.matrix}
\title{Generate a symmetric Q matrix}
\usage{
symmetric.Q.matrix(K)
}
\arguments{
\item{K}{Integer. The number of states.}
}
\value{
A K x K numeric matrix representing the symmetric Q matrix.
}
\description{
Creates a K x K symmetric rate matrix (Q matrix) with equal transition rates between states.
The diagonal elements are set such that each row sums to zero.
}
\examples{
symmetric.Q.matrix(4)
}
