% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write.recon.matrix}
\alias{write.recon.matrix}
\title{Write reconstructed character matrix to file}
\usage{
write.recon.matrix(data, file = NULL, all)
}
\arguments{
\item{data}{Morpho object}

\item{file}{File name}

\item{all}{Logical. If TRUE, the sequences of all sampled ancestors in the
reconstructed tree are written, not just the first on each lineage.
Default FALSE.}
}
\value{
No return value, called for its side effect of writing data to a file.
}
\description{
Write the character matrix for the reconstructed tree to a nexus file
}
\examples{
data(morpho_data)
tmp <- tempfile(fileext = ".nex")
write.recon.matrix(data = morpho_data, file = tmp)

}
