% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write.recon.tree}
\alias{write.recon.tree}
\title{Write reconstructed tree to file}
\usage{
write.recon.tree(data = NULL, file = NULL, all = TRUE, reconstructed = TRUE)
}
\arguments{
\item{data}{Morpho object}

\item{file}{File name}
}
\value{
No return value, called for its side effect of writing data to a file.
}
\description{
Write the reconstructed tree to Newick string
}
\examples{
data(morpho_data)
tmp <- tempfile(fileext = ".tre")
write.recon.tree(data = morpho_data, file = tmp)


}
