% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write.tsv}
\alias{write.tsv}
\title{Write the taxa ages}
\usage{
write.tsv(data, file, uncertainty = 0, all)
}
\arguments{
\item{data}{Morpho object}

\item{file}{File name}

\item{uncertainty}{Numeric. Adds uncertainty to fossil ages in the morpho object.
The ages in the object are point estimates by default; setting `uncertainty`
will create an age range of ± this value (in millions of years).}

\item{all}{Logical. If TRUE, also write the ages of internal nodes.
Internal node ages are written as point values (no uncertainty). Default FALSE.}
}
\value{
No return value, called for its side effect of writing data to a file.
}
\description{
Writes the ages of the specimens in the true tree to a file. The tsv format used
here is directly compatible with RevBayes
}
\examples{
data(morpho_data)
tmp <- tempfile(fileext = ".tsv")
write.tsv(data = morpho_data, file = tmp)

}
