% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{morphsim_fossilsim}
\alias{morphsim_fossilsim}
\title{Match sampled ancestor labels}
\usage{
morphsim_fossilsim(data = NULL)
}
\arguments{
\item{data}{Morpho object containing fossils}
}
\value{
A character matrix mapping sampled ancestor labels between the naming
conventions used by \code{Morphsim} and \code{Fossilsim}
}
\description{
Match the sampled ancestor labels from \code{Morphsim} and \code{Fossilsim}
}
\examples{
data(morpho_data)
morphsim_fossilsim <- function(data = morpho_data)

}
