% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write.morpho}
\alias{write.morpho}
\title{Write morpho data to file}
\usage{
write.morpho(
  data,
  file,
  type = "tree",
  all = FALSE,
  reconstructed = FALSE,
  uncertainty = 0
)
}
\arguments{
\item{data}{A morpho object}

\item{file}{File name}

\item{type}{type to write: "tree", "matrix", or "ages"}

\item{all}{Logical. If TRUE, write data for all sampled specimens. Applies to
`type = "matrix"` and `type = "ages"` Default FALSE.}

\item{reconstructed}{If TRUE, write the reconstructed version. Default FALSE.}

\item{uncertainty}{Numeric. Age uncertainty for fossil ages. Default 0.}
}
\value{
No return value, called for its side effect of writing data to a file.
}
\description{
Export components of a morpho object to file. Can write trees, character
matrices, or fossil ages in various formats.
}
\examples{
data(morpho_data)
tmp <- tempfile(fileext = ".tre")
write.morpho(morpho_data, file = tmp, type = "tree")

}
